requirements-core.txt
fastapi~=0.42.0
uvicorn~=0.9.0 
etcd3~=0.10.0
starlette-prometheus~=0.5.0
func-timeout~=4.3.5
sqlalchemy~=1.3.11
psycopg2-binary~=2.8.4
requests~=2.22.0
json_log_formatter~=0.2.0
google-cloud-pubsub
